+++ .eslintrc.json
@@ -3,6 +3,7 @@
     "browser": true,
     "es6": true
   },
+  "plugins": ["vue"],
   "extends": "eslint:recommended",
   "rules": {
     "eqeqeq": [
