+++ nuxt.config.js
@@ -26,6 +26,8 @@ export default {
 
   // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
   plugins: [
+    '~/plugins/speech',
+    '~/plugins/recognition'
   ],
 
   // Auto import components: https://go.nuxtjs.dev/config-components
